decrating: collapse dual implementation for environment (preserve shim logic)
Eliminates the dual-path pattern for the environment subsystem:
- Deletes the stale in-tree `environment.rs` (had only
  `collect_environment_fingerprint` via std::process::Command calls).
- Promotes `environment_micro.rs` to canonical `environment.rs`.
- The shim's added logic is preserved: it delegates to
  `shipper_environment::EnvironmentInfo::collect` (reusing CI detection,
  env-var capture, etc.) and normalizes the `cargo`/`rustc` version
  strings via `normalize_version` before mapping them into
  `EnvironmentFingerprint`. It also provides a graceful fallback when
  collection fails.
- Removes `micro-environment` feature from shipper and shipper-cli
  Cargo.toml, makes `shipper-environment` a non-optional dep, drops the
  cross-reference from `micro-state` and `micro-all`.
