+++ crates/shipper-cli/Cargo.toml
@@ -33,7 +33,6 @@ workspace = true
 default = ["micro-all"]
 micro-git = ["shipper/micro-git"]
 micro-encrypt = ["shipper/micro-encrypt"]
-micro-environment = ["shipper/micro-environment"]
 micro-storage = ["shipper/micro-storage"]
 micro-cargo = ["shipper/micro-cargo"]
 micro-plan = ["shipper/micro-plan"]
