+++ crates/shipper/Cargo.toml
@@ -19,7 +19,7 @@ shipper-encrypt = { path = "../shipper-encrypt" }
 shipper-git = { path = "../shipper-git", optional = true }
 shipper-events = { path = "../shipper-events" }
 shipper-execution-core = { path = "../shipper-execution-core" }
-shipper-environment = { path = "../shipper-environment", optional = true }
+shipper-environment = { path = "../shipper-environment" }
 shipper-storage = { path = "../shipper-storage", optional = true }
 shipper-cargo = { path = "../shipper-cargo", optional = true }
 shipper-engine-parallel = { path = "../shipper-engine-parallel", optional = true }
@@ -63,7 +63,6 @@ sha2 = "0.11"
 [features]
 micro-git = ["dep:shipper-git"]
 micro-encrypt = []
-micro-environment = ["dep:shipper-environment", "micro-types"]
 micro-storage = ["dep:shipper-storage"]
 micro-cargo = ["dep:shipper-cargo"]
 micro-plan = ["dep:shipper-plan", "micro-types"]
@@ -71,7 +70,7 @@ micro-webhook = []
 micro-types = ["micro-webhook", "micro-encrypt", "micro-config"]
 micro-config = ["micro-types", "micro-storage"]
 micro-parallel = ["micro-types", "micro-plan", "dep:shipper-engine-parallel"]
-micro-all = ["micro-git", "micro-encrypt", "micro-environment", "micro-storage", "micro-cargo", "micro-plan", "micro-webhook", "micro-types", "micro-config", "micro-parallel"]
+micro-all = ["micro-git", "micro-encrypt", "micro-storage", "micro-cargo", "micro-plan", "micro-webhook", "micro-types", "micro-config", "micro-parallel"]
 
 [lints]
 workspace = true
